% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calculer la distribution des âges}
\usage{
calcul_distribution_age(df, ref_date = Sys.Date(), ...)
}
\arguments{
\item{df}{Un data.frame contenant une colonne \verb{Date de naissance} et
une colonne \verb{Libellé de la commune} ou \verb{Libellé du département}.}

\item{...}{Paramètres supplémentaires (non utilisés directement dans cette fonction).}
}
\value{
Un tibble contenant le nom de la commune ou du département, ainsi que
les statistiques d'âge : minimum, premier quartile (Q1), médiane, moyenne,
troisième quartile (Q3) et maximum.
}
\description{
Cette fonction calcule la distribution des âges des individus d'un data.frame
en renvoyant des statistiques descriptives : minimum, quartiles, médiane, moyenne et maximum.
}
\keyword{internal}
