% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Compter le nombre d'adjoints}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un data.frame contenant une colonne \verb{Libellé de la fonction} (type \code{character}).}
}
\value{
Un entier représentant le nombre d'occurrences du mot "adjoint".
}
\description{
Cette fonction compte le nombre d'occurrences du mot "adjoint"
dans la colonne \verb{Libellé de la fonction} d'un data.frame.
}
\keyword{internal}
