% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_elus.R
\name{compter_nombre_d_elus}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élus}
\usage{
compter_nombre_d_elus(df)
}
\arguments{
\item{df}{Un data.frame contenant les colonnes \verb{Nom de l'élu}, \verb{Prénom de l'élu}, et \verb{Date de naissance} (type \code{character} pour les deux premières et \code{Date} pour la dernière).}
}
\value{
Un entier représentant le nombre d'élus distincts dans le data.frame.
}
\description{
Cette fonction compte le nombre d'élus distincts dans un data.frame en se basant
sur les colonnes \verb{Nom de l'élu}, \verb{Prénom de l'élu}, et \verb{Date de naissance}.
}
\keyword{internal}
