% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_code_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Visualiser la répartition des élus par catégorie socio-professionnelle}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un data.frame contenant une colonne \verb{Code de la catégorie socio-professionnelle}.}
}
\value{
Un objet ggplot affichant un diagramme en barres de la distribution des élus
par catégorie socio-professionnelle.
}
\description{
Cette fonction génère un graphique en barres représentant la répartition
des élus selon leur catégorie socio-professionnelle.
}
\keyword{internal}
