% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valider_schema.R
\name{valider_schema}
\alias{valider_schema}
\title{Valider le schéma des données d'un data.frame}
\usage{
valider_schema(df)
}
\arguments{
\item{df}{Un data.frame, contenant des informations sur les élus, qui doit avoir au minimum les colonnes suivantes :
\verb{Code du département}, \verb{Libellé du département}, \verb{Code de la commune}, \verb{Libellé de la commune},
\verb{Nom de l'élu}, \verb{Prénom de l'élu}, \verb{Date de naissance}, \verb{Code de la catégorie socio-professionnelle},
\verb{Libellé de la fonction}.}
}
\value{
Aucun retour explicite. Un message ou un avertissement est affiché si le schéma n’est pas respecté.
}
\description{
Cette fonction vérifie que les colonnes du data.frame sont conformes à un schéma prédéfini. Le schéma attendu inclut des colonnes spécifiques
liées aux informations sur les élus.
}
\keyword{internal}
